chore: add unit tests
