+++ pkg/suggest/rules_test.go
@@ -13,43 +13,3 @@
 
 package suggest
 
-import (
-	"errors"
-	"github.com/stretchr/testify/assert"
-	"testing"
-)
-
-func TestUserFriendlyError(t *testing.T) {
-	tests := []struct {
-		name     string
-		input    error
-		expected string
-	}{
-		{
-			name:     "empty",
-			input:    errors.New(""),
-			expected: "",
-		},
-		{
-			name:  "yaml errors have heading",
-			input: errors.New("yaml: some random error"),
-			expected: `Your okteto manifest is not valid, please check the following errors:
-yaml: some random error.
-    Check out the okteto manifest docs at: https://www.okteto.com/docs/reference/manifest`,
-		},
-		{
-			name:  "yaml errors have heading",
-			input: errors.New("yaml: some random error"),
-			expected: `Your okteto manifest is not valid, please check the following errors:
-yaml: some random error.
-    Check out the okteto manifest docs at: https://www.okteto.com/docs/reference/manifest`,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := UserFriendlyError(tt.input)
-			assert.Equal(t, tt.expected, err.Error())
-		})
-	}
-}
