+++ pkg/suggest/suggest_test.go
@@ -101,3 +101,38 @@ func TestErrorSuggestion(t *testing.T) {
 		})
 	}
 }
+
+func TestUserFriendlyError(t *testing.T) {
+	tests := []struct {
+		name     string
+		input    error
+		expected string
+	}{
+		{
+			name:     "empty",
+			input:    errors.New(""),
+			expected: "",
+		},
+		{
+			name:  "yaml errors with heading and link to docs",
+			input: errors.New("yaml: some random error"),
+			expected: `Your okteto manifest is not valid, please check the following errors:
+yaml: some random error.
+    Check out the okteto manifest docs at: https://www.okteto.com/docs/reference/manifest`,
+		},
+		{
+			name:  "yaml errors with heading and link to docs",
+			input: errors.New("yaml: unmarshal errors:\n  line 4: field contex not found in type model.buildInfoRaw"),
+			expected: `Your okteto manifest is not valid, please check the following errors:
+yaml: some random error.
+    Check out the okteto manifest docs at: https://www.okteto.com/docs/reference/manifest`,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			err := UserFriendlyError(tt.input)
+			assert.Equal(t, tt.expected, err.Error())
+		})
+	}
+}
